Extract studio page test harness
